Fix overflow

+++ source/advanced_two_hop.h
@@ -22,7 +22,7 @@ class AdvancedTwoHopIndex {
         
         bool is_temporal_path;
         int k = 0;
-        int visited_paths = 0;
+        long long visited_paths = 0;
         int* order;
 
         int size();
